# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AnimeSearchResult {
  title: string;
  url: string;
  slug: string;
  image: string;
}

export interface SearchResponse {
  success: boolean;
  query: string;
  results: AnimeSearchResult[];
}

export interface AnimeEpisode {
  id: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  season?: number | null;
  /** @nullable */
  number?: number | null;
  /** @nullable */
  url?: string | null;
}

export interface AnimeSeason {
  number: number;
  episodes: AnimeEpisode[];
}

export interface SeriesData {
  title: string;
  slug: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  description?: string | null;
  genres?: string[];
  seasons?: AnimeSeason[];
  /** @nullable */
  total_episodes?: number | null;
}

export interface SeriesResponse {
  success: boolean;
  data: SeriesData;
}

export interface EpisodeData {
  /** @nullable */
  video_player?: string | null;
  /** @nullable */
  m3u8_link?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  next_episode_id?: string | null;
  /** @nullable */
  prev_episode_id?: string | null;
}

export interface EpisodeResponse {
  success: boolean;
  data: EpisodeData;
}

export type SearchAnimeParams = {
q: string;
};