# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AnimeSalt streaming API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: anime
    description: Anime streaming operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /anime/search:
    get:
      operationId: searchAnime
      tags: [anime]
      summary: Search anime by query
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"

  /anime/series/{slug}:
    get:
      operationId: getAnimeSeries
      tags: [anime]
      summary: Get anime series details
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Series details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /anime/episode/{episodeId}:
    get:
      operationId: getAnimeEpisode
      tags: [anime]
      summary: Get episode streaming data
      parameters:
        - in: path
          name: episodeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Episode streaming data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AnimeSearchResult:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        slug:
          type: string
        image:
          type: string
      required:
        - title
        - url
        - slug
        - image

    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/AnimeSearchResult"
      required:
        - success
        - query
        - results

    AnimeEpisode:
      type: object
      properties:
        id:
          type: string
        title:
          type: ["string", "null"]
        season:
          type: ["integer", "null"]
        number:
          type: ["integer", "null"]
        url:
          type: ["string", "null"]
      required:
        - id

    AnimeSeason:
      type: object
      properties:
        number:
          type: integer
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/AnimeEpisode"
      required:
        - number
        - episodes

    SeriesData:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        image:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        seasons:
          type: array
          items:
            $ref: "#/components/schemas/AnimeSeason"
        total_episodes:
          type: ["integer", "null"]
      required:
        - title
        - slug

    SeriesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/SeriesData"
      required:
        - success
        - data

    EpisodeData:
      type: object
      properties:
        video_player:
          type: ["string", "null"]
        m3u8_link:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        next_episode_id:
          type: ["string", "null"]
        prev_episode_id:
          type: ["string", "null"]
      required: []

    EpisodeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/EpisodeData"
      required:
        - success
        - data