# lib/api-zod/src/generated/types/animeEpisode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnimeEpisode {
  id: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  season?: number | null;
  /** @nullable */
  number?: number | null;
  /** @nullable */
  url?: string | null;
}