# lib/api-zod/src/generated/types/animeSearchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnimeSearchResult {
  title: string;
  url: string;
  slug: string;
  image: string;
}