# lib/api-zod/src/generated/types/animeSeason.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeEpisode } from './animeEpisode';

export interface AnimeSeason {
  number: number;
  episodes: AnimeEpisode[];
}