# lib/api-zod/src/generated/types/seriesData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeSeason } from './animeSeason';

export interface SeriesData {
  title: string;
  slug: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  description?: string | null;
  genres?: string[];
  seasons?: AnimeSeason[];
  /** @nullable */
  total_episodes?: number | null;
}